requirements.txt
moviepy~=1.0.3
pyt2s~=0.0.4
pydub~=0.25.1
praw~=7.7.1
python-dotenv~=1.0.1
opencv-python~=1.23.4
transformers~=4.43.3
playwright~=1.45.1
asyncpraw~=7.7.1
httplib2~=0.22.0
google-api-python-client~=2.139.0
pytz~=2024.1